# services/api/openapi.json
{
  "components": {
    "schemas": {
      "Authorization": {
        "description": "What the console brings back from Keycloak: the code and the verifier it kept.",
        "properties": {
          "code": {
            "maxLength": 4096,
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "code_verifier": {
            "pattern": "^[A-Za-z0-9\\-._~]{43,128}$",
            "title": "Code Verifier",
            "type": "string"
          },
          "redirect_uri": {
            "maxLength": 2000,
            "pattern": "^https?://",
            "title": "Redirect Uri",
            "type": "string"
          }
        },
        "required": [
          "code",
          "code_verifier",
          "redirect_uri"
        ],
        "title": "Authorization",
        "type": "object"
      },
      "Exercise": {
        "properties": {
          "answered_at": {
            "description": "when the client answered it",
            "format": "date-time",
            "title": "Answered At",
            "type": "string"
          },
          "requested_at": {
            "description": "when the client received the request",
            "format": "date-time",
            "title": "Requested At",
            "type": "string"
          },
          "right": {
            "default": "erasure",
            "description": "the right exercised",
            "maxLength": 40,
            "minLength": 1,
            "title": "Right",
            "type": "string"
          }
        },
        "required": [
          "requested_at",
          "answered_at"
        ],
        "title": "Exercise",
        "type": "object"
      },
...
      "Problem": {
        "description": "An error as RFC 9457 describes it: the same shape whatever went wrong.",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "what happened in this occurrence",
            "title": "Detail"
          },
          "instance": {
            "description": "path where it happened",
            "title": "Instance",
            "type": "string"
          },
          "status": {
            "description": "HTTP status code",
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "description": "short human summary, the same for every occurrence",
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "description": "identifier of the kind of problem",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "instance"
        ],
        "title": "Problem",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "access token issued by the appliance Keycloak",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Campaigns, inventory, findings and evidence of the ARGOS appliance. Listings paginate by opaque cursor, errors follow RFC 9457 and every POST that creates accepts an Idempotency-Key.",
    "title": "ARGOS API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "logout_api_v1_auth_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "no valid token"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "the role does not allow it"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "it does not exist"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "the state does not allow it"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "invalid body"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "not configured yet"
          }
        },
        "summary": "Close the console session: revoke the refresh token and delete its cookie",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "refresh_api_v1_auth_refresh_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "no valid token"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "the role does not allow it"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "it does not exist"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "the state does not allow it"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "invalid body"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "not configured yet"
          }
        },
        "summary": "Exchange the refresh cookie for an access token",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/session": {
      "post": {
        "operationId": "open_session_api_v1_auth_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Authorization"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "no valid token"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "the role does not allow it"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "it does not exist"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "the state does not allow it"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "invalid body"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "not configured yet"
          }
        },
        "summary": "Open the console session from an authorization code (PKCE)",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/synthetic/{injection_id}/confirm-exercise": {
      "post": {
        "operationId": "exercised_api_v1_synthetic__injection_id__confirm_exercise_post",
        "parameters": [
          {
            "in": "path",
            "name": "injection_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Injection Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Exercise"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Exercised Api V1 Synthetic  Injection Id  Confirm Exercise Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "no valid token"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "the role does not allow it"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "it does not exist"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "the state does not allow it"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "invalid body"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "not configured yet"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "The client says the subject exercised a right",
        "tags": [
          "synthetic"
        ]
      }
    },
    "/api/v1/synthetic/{injection_id}/confirm-injection": {
      "post": {
        "operationId": "injected_api_v1_synthetic__injection_id__confirm_injection_post",
        "parameters": [
          {
            "in": "path",
            "name": "injection_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Injection Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Injected Api V1 Synthetic  Injection Id  Confirm Injection Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "no valid token"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "the role does not allow it"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "it does not exist"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "the state does not allow it"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "invalid body"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "not configured yet"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "The client says it injected the subject at the authorised point",
        "tags": [
          "synthetic"
        ]
      }
    },
    "/api/v1/synthetic/{injection_id}/confirm-revert": {
      "post": {
        "operationId": "reverted_api_v1_synthetic__injection_id__confirm_revert_post",
        "parameters": [
          {
            "in": "path",
            "name": "injection_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Injection Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Reverted Api V1 Synthetic  Injection Id  Confirm Revert Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "no valid token"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "the role does not allow it"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "it does not exist"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "the state does not allow it"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "invalid body"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "not configured yet"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "The client says it put the source back as it was",
        "tags": [
          "synthetic"
        ]
      }
    },
...
  }
}